refactor: delete commited code, change folder name to kebab
