+++ src/stories/button.stories.ts
@@ -1,7 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import ButtonMUI from '../shared/ui/ButtonMUI';
+import ButtonMUI from '../shared/ui/Button-mui';
 
-// More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 const meta = {
 	title: 'Example/Button',
 	component: ButtonMUI,
@@ -11,16 +10,11 @@ const meta = {
 	},
 	// This component will have an automatically generated Autodocs entry: https://storybook.js.org/docs/react/writing-docs/autodocs
 	tags: ['autodocs'],
-	// More on argTypes: https://storybook.js.org/docs/react/api/argtypes
-	// argTypes: {
-	// 	backgroundColor: { control: 'color' },
-	// },
 } satisfies Meta<typeof ButtonMUI>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
 
-// More on writing stories with args: https://storybook.js.org/docs/react/writing-stories/args
 export const Primary: Story = {
 	args: {
 		text: 'Войти',
